Refactor: Table head컴포넌트 컴파운드 패턴으로 변경
