+++ src/types/report.dto.ts
